Tweaking waveform viewer

+++ Editor/Windows/WaveformWindow.cs
@@ -8,87 +8,83 @@ namespace Cinemachine.Editor
 {
     public class WaveformWindow : EditorWindow
     {
+        WaveformGenerator mWaveformGenerator;
         Texture2D mScreenshot;
         string mScreenshotFilename;
-        WaveformGenerator mWaveformGenerator;
 
         [MenuItem("Window/Waveform Monitor")]
         private static void OpenWindow()
         {
-            WaveformWindow window = EditorWindow.GetWindow<WaveformWindow>(true);
+            WaveformWindow window = EditorWindow.GetWindow<WaveformWindow>(false);
             window.autoRepaintOnSceneChange = true;
-            window.position = new Rect(100, 100, 400, 400);
+            //window.position = new Rect(100, 100, 400, 400);
             window.Show(true);
         }
         
         private void OnEnable()
         {
             titleContent = new GUIContent("Waveform", CinemachineSettings.CinemachineLogoTexture);
+            mWaveformGenerator = new WaveformGenerator();
+
             mScreenshotFilename = Path.GetFullPath(FileUtil.GetUniqueTempPathInProject() + ".png");
             ScreenCapture.CaptureScreenshot(mScreenshotFilename);
             EditorApplication.update += UpdateScreenshot;
-            mWaveformGenerator = new WaveformGenerator();
         }
 
         private void OnDisable()
         {
             EditorApplication.update -= UpdateScreenshot;
-            if (File.Exists(mScreenshotFilename))
+            if (!string.IsNullOrEmpty(mScreenshotFilename) && File.Exists(mScreenshotFilename))
                 File.Delete(mScreenshotFilename);
             mScreenshotFilename = null;
             mWaveformGenerator.DestroyBuffers();
+            if (mScreenshot != null)
+                DestroyImmediate(mScreenshot);
+            mScreenshot = null;
         }
 
         private void OnGUI()
         {
-            if (mWaveformGenerator.Result != null)
-            {
-                Rect rect = EditorGUILayout.GetControlRect(true);
-                EditorGUIUtility.labelWidth /= 2;
-                mWaveformGenerator.m_Exposure = EditorGUI.Slider(
-                    rect, "Exposure", mWaveformGenerator.m_Exposure, 0.01f, 2);
-                EditorGUIUtility.labelWidth *= 2;
-                rect.y += rect.height;
-                rect.height = position.height - rect.height;
-                GUI.DrawTexture(rect, mWaveformGenerator.Result, ScaleMode.StretchToFill);
-            }
+            Rect rect = EditorGUILayout.GetControlRect(true);
+            EditorGUIUtility.labelWidth /= 2;
+            EditorGUI.BeginChangeCheck();
+            mWaveformGenerator.m_Exposure = EditorGUI.Slider(
+                rect, "Exposure", mWaveformGenerator.m_Exposure, 0.01f, 2);
+            if (EditorGUI.EndChangeCheck())
+                UnityEditorInternal.InternalEditorUtility.RepaintAllViews();
+            EditorGUIUtility.labelWidth *= 2;
+            rect.y += rect.height;
+            rect.height = position.height - rect.height;
+            var tex = mWaveformGenerator.Result;
+            if (tex != null)
+                GUI.DrawTexture(rect, tex);
         }
 
         float mLastUpdateTime = 0;
         private void UpdateScreenshot()
         {
-            if (mScreenshotFilename == null)
-                return;
-
             // Don't do this too often
             float now = Time.time;
             if (mScreenshot != null && now - mLastUpdateTime < 0.1f)
                 return;
 
             mLastUpdateTime = now;
-            if (mScreenshot != null)
-                Object.DestroyImmediate(mScreenshot);
-            mScreenshot = null;
-#if false // GML temp hack CaptureScreenshotAsTexture() is broken in Unity
-            mScreenshot = ScreenCapture.CaptureScreenshotAsTexture();
-#else
-            if (File.Exists(mScreenshotFilename))
+            if (!string.IsNullOrEmpty(mScreenshotFilename) && File.Exists(mScreenshotFilename))
             {
                 byte[] fileData = File.ReadAllBytes(mScreenshotFilename);
-                mScreenshot = new Texture2D(2, 2);
+                if (mScreenshot == null)
+                    mScreenshot = new Texture2D(2, 2);
                 mScreenshot.LoadImage(fileData); // this will auto-resize the texture dimensions.
                 mWaveformGenerator.RenderWaveform(mScreenshot);
 
                 // Capture the next one
                 ScreenCapture.CaptureScreenshot(mScreenshotFilename);
-                Repaint();
             }
-#endif
         }
 
         class WaveformGenerator
         {
-            public float m_Exposure = 0.12f;
+            public float m_Exposure = 0.2f;
 
             RenderTexture mOutput;
             ComputeBuffer mData;
@@ -101,7 +97,28 @@ namespace Cinemachine.Editor
             MaterialPropertyBlock mWaveformProperties;
             Material mWaveformMaterial;
             CommandBuffer mCmd;
+
             static Mesh sFullscreenTriangle;
+            static Mesh FullscreenTriangle 
+            { 
+                get 
+                {
+                    if (sFullscreenTriangle == null)
+                    {
+                        sFullscreenTriangle = new Mesh { name = "Fullscreen Triangle" };
+                        sFullscreenTriangle.SetVertices(new List<Vector3>
+                        {
+                            new Vector3(-1f, -1f, 0f),
+                            new Vector3(-1f,  3f, 0f),
+                            new Vector3( 3f, -1f, 0f)
+                        });
+                        sFullscreenTriangle.SetIndices(
+                            new [] { 0, 1, 2 }, MeshTopology.Triangles, 0, false);
+                        sFullscreenTriangle.UploadMeshData(false);
+                    }
+                    return sFullscreenTriangle;
+                } 
+            }
           
             public WaveformGenerator()
             {
@@ -113,20 +130,6 @@ namespace Cinemachine.Editor
                     hideFlags = HideFlags.DontSave
                 };
                 mCmd = new CommandBuffer();
-
-                if (sFullscreenTriangle == null)
-                {
-                    sFullscreenTriangle = new Mesh { name = "Fullscreen Triangle" };
-                    sFullscreenTriangle.SetVertices(new List<Vector3>
-                    {
-                        new Vector3(-1f, -1f, 0f),
-                        new Vector3(-1f,  3f, 0f),
-                        new Vector3( 3f, -1f, 0f)
-                    });
-                    sFullscreenTriangle.SetIndices(
-                        new [] { 0, 1, 2 }, MeshTopology.Triangles, 0, false);
-                    sFullscreenTriangle.UploadMeshData(false);
-                }
             }
 
             void CreateBuffers(int width, int height)
@@ -145,11 +148,9 @@ namespace Cinemachine.Editor
                     mOutput.Create();
                 }
 
-                int count = width * height;
+                int count = Mathf.CeilToInt(width / (float)mThreadGroupSizeX) * mThreadGroupSizeX * height;
                 if (mData == null)
-                {
                     mData = new ComputeBuffer(count, sizeof(uint) << 2);
-                }
                 else if (mData.count < count)
                 {
                     mData.Release();
@@ -170,21 +171,25 @@ namespace Cinemachine.Editor
 
             public void RenderWaveform(Texture2D source)
             {
+                if (mWaveformMaterial == null)
+                    return;
+
                 int width = source.width;
                 int height = source.height;
+                int histogramResolution = 256;
 
-                CreateBuffers(width, height);
-
-                mThreadGroupSizeX = 16;
                 mThreadGroupSize = 256;
+                mThreadGroupSizeX = 16;
                 mThreadGroupSizeY = 16;
+                CreateBuffers(width, histogramResolution);
 
                 mCmd.Clear();
                 mCmd.BeginSample("CMWaveform");
 
                 var parameters = new Vector4(
                     width, height, 
-                    QualitySettings.activeColorSpace == ColorSpace.Linear ? 1 : 0, 0f);
+                    QualitySettings.activeColorSpace == ColorSpace.Linear ? 1 : 0, 
+                    histogramResolution);
 
                 // Clear the buffer on every frame
                 int kernel = mWaveformCompute.FindKernel("KCMWaveformClear");
@@ -192,7 +197,7 @@ namespace Cinemachine.Editor
                 mCmd.SetComputeVectorParam(mWaveformCompute, "_Params", parameters);
                 mCmd.DispatchCompute(mWaveformCompute, kernel, 
                     Mathf.CeilToInt(width / (float)mThreadGroupSizeX), 
-                    Mathf.CeilToInt(height / (float)mThreadGroupSizeY), 1);
+                    Mathf.CeilToInt(histogramResolution / (float)mThreadGroupSizeY), 1);
 
                 // Gather all pixels and fill in our waveform
                 kernel = mWaveformCompute.FindKernel("KCMWaveformGather");
@@ -203,14 +208,17 @@ namespace Cinemachine.Editor
                     Mathf.CeilToInt(height / (float)mThreadGroupSize), 1);
 
                 // Generate the waveform texture
+                float exposure = Mathf.Max(0f, m_Exposure);
+                exposure *= (float)histogramResolution / height;
                 mWaveformProperties.SetVector(Shader.PropertyToID("_Params"), 
-                    new Vector4(width, height, Mathf.Max(0f, m_Exposure), 0f));
+                    new Vector4(width, histogramResolution, exposure, 0f));
                 mWaveformProperties.SetBuffer(Shader.PropertyToID("_WaveformBuffer"), mData);
                 mCmd.SetRenderTarget(mOutput);
                 mCmd.DrawMesh(
-                    sFullscreenTriangle, Matrix4x4.identity, 
+                    FullscreenTriangle, Matrix4x4.identity, 
                     mWaveformMaterial, 0, 0, mWaveformProperties);
                 mCmd.EndSample("CMWaveform");
+
                 Graphics.ExecuteCommandBuffer(mCmd);
             }
         }
